requirements.txt
fastapi
uvicorn
pillow
pytesseract
paddleocr
ultralytics
opencv-python
numpy
PyMuPDF
